pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "game-map-render"
version = "0.1.0"
description = "Generate beautiful tile images for game maps using local Stable Diffusion"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "torch>=2.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "safetensors",
    "Pillow>=10.0",
    "click>=8.0",
    "numpy",
]

[project.scripts]
game-map-render = "src.main:generate"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
